EnemyRange 버그수정, EnemyBullet 데미지판정 제작해야함

+++ Assets/02.Scripts/Bullets/Bullets.cs
@@ -5,5 +5,5 @@ using UnityEngine;
 public abstract class Bullets : MonoBehaviour
 {
     public int damage;
-    public int knockbackSize;
+    public float knockbackSize;
 }
